squash! Refactor the functions that run over every event.
Fixes suggested by Dave Hart:

Add typedefs for the callback types.

Name fewer things "ctx".

Also, when this commit message gets squashed in, say:

Adds an implementation of Floyd's tortoise-and-hare algorithm to check
for circularity in TAILQs and LISTs, to avoid the abuse of flags that
event_base_assert_ok() was doing before.